Splits modules into separate files
