+++ lib/console_in_out.ex
@@ -1,0 +1,13 @@
+defmodule ConsoleInOut do
+  def print(output) do
+    IO.puts(output)
+  end
+
+  def read_move do
+    IO.gets "\nEnter a number to make your move: "
+  end
+
+  def read_input(prompt) do
+    IO.gets prompt
+  end
+end
